fix: Payload crear documento venta

+++ app/models/pydantic/world_office/facturacion.py
@@ -12,9 +12,9 @@ class WOReglone(Base):
     idInventario: int = 0
     unidadMedida: str = ''
     cantidad: int = 0
-    valorUnitario: float = 0
+    valorUnitario: int = 0
     idBodega: int = 0
-    porDescuento: float = 0
+    porDescuento: int = 0
     idCentroCosto: int | None = None
